backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
python-dotenv==1.0.0
django-filter==23.5
